Es können nun AnchorPoints hinzugefügt werden, aber noch nicht verschoben werden

+++ DigitSim/src/connection/AnchorPoint.java
@@ -13,9 +13,9 @@ import general.Vector2i;
  */
 public class AnchorPoint {
     Vector2i coords;
-    int index;
+    double index;
 
-    public AnchorPoint(int index, Vector2i coords) {
+    public AnchorPoint(double index, Vector2i coords) {
         this.coords = coords;
         this.index = index;
     }
@@ -37,11 +37,11 @@ public class AnchorPoint {
         this.coords = coords;
     }
 
-    public int getIndex() {
+    public double getIndex() {
         return index;
     }
 
-    public void setIndex(int index) {
+    public void setIndex(double index) {
         this.index = index;
     }
